+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v6.0.2-1.19.4] - 2023-05-20
+### Changed
+- Implemented `RenderGuiElementEvents` for item name
+- Moved `SpawnDataMob` implementation to Fabric, the native `Mob::getSpawnType` method is now used on Forge
+- Allow more context for `RenderGuiElementEvents` to ease implementations on different mod loaders 
+### Fixed
+- Fixed `ClassCastException` for after events in `ScreenMouseEvents`
+
 ## [v6.0.1-1.19.4] - 2023-05-19
 ### Added
 - Added events for when the player is being cloned and for after respawning